Implemented enabling movement
Inside the PathNavigator script I've added another coroutine called DelayEnableMoving and gets called in the on trigger enter function.  Inside the function there is a short wait time set by the variable enableMovementCooldown before executing the rest of the code within the function. After that time the isMoving bool gets set to false and the progress float gets set back to zero.